requirements.txt
tensorflow
keras
streamlit 
kaggle 
numpy 
pandas
matplotlib 
seaborn
pillow 
scikit-learn